Fix the error of computing fib(93) and larger
In this implenmentation of computing fibonnacci number, bignum struct
is used for store the number larger than uint. And the result was
printed in string type.

Adapted the file 'client.c' and 'fibdrv.c' to make the fibonnacci
implenment work in kernel driver.

'expected.txt' was fixed to pass the test.